fix: 🐛 put styles in seprate file and optimize the code
