+++ src/components/Button/styles.js
@@ -1,7 +1,7 @@
-const style = {
+const styles = {
   padding: '15px',
   fontSize: '16px',
   margin: '5px',
 };
 
-export default style;
+export default styles;
