+++ src/components/Progress/styles.js
@@ -1,0 +1,7 @@
+const styles = theme => ({
+  progress: {
+    margin: theme.spacing.unit * 0.2,
+  },
+})
+
+export default styles;
