+++ src/contexts/SnackBarProvider/styles.js
@@ -1,0 +1,26 @@
+import green from '@material-ui/core/colors/green';
+
+const styles = theme => ({
+  close: {
+    padding: theme.spacing.unit / 2,
+  },
+  success: {
+    backgroundColor: green[600],
+  },
+  error: {
+    backgroundColor: theme.palette.error.dark,
+  },
+  message: {
+    display: 'flex',
+    alignItems: 'center',
+  },
+  iconVariant: {
+    opacity: 0.9,
+    marginRight: theme.spacing.unit,
+  },
+  icon: {
+    fontSize: 20,
+  },
+});
+
+export default styles;
