+++ src/layouts/components/Footer/Footer.jsx
@@ -2,13 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { withStyles } from '@material-ui/core/styles';
 import Typography from '@material-ui/core/Typography';
-
-const styles = theme => ({
-  footer: {
-    marginTop: theme.spacing.unit * 8,
-    padding: `${theme.spacing.unit * 6}px 0`,
-  },
-});
+import styles from './styles';
 
 function Footer(props) {
   const { classes } = props;
