+++ src/layouts/components/Footer/styles.js
@@ -1,0 +1,8 @@
+const styles = theme => ({
+  footer: {
+    marginTop: theme.spacing.unit * 8,
+    padding: `${theme.spacing.unit * 6}px 0`,
+  },
+});
+
+export default styles;
