+++ src/pages/NoMatch/styles.js
@@ -1,0 +1,8 @@
+const styles = theme => ({
+  div: {
+    marginTop: theme.spacing.unit,
+    padding: `${theme.spacing.unit * 6}px 0`,
+  },
+});
+
+export default styles;
