+++ src/pages/Trainee/components/EditDialog/styles.js
@@ -1,0 +1,12 @@
+const styles = theme => ({
+  textField: {
+    marginLeft: theme.spacing.unit,
+    marginRight: theme.spacing.unit,
+  },
+  error: {
+    color: 'red',
+    margin: 10,
+  },
+});
+
+export default styles;
