+++ src/pages/components/Navbar/Navbar.jsx
@@ -5,22 +5,8 @@ import AppBar from '@material-ui/core/AppBar';
 import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
 import Button from '@material-ui/core/Button';
+import styles from './styles';
 
-const styles = {
-  root: {
-    flexGrow: 1,
-  },
-  grow: {
-    flexGrow: 1,
-  },
-  menuButton: {
-    marginLeft: -12,
-    marginRight: 20,
-  },
-  loginButton: {
-    marginLeft: '30px',
-  },
-};
 
 function Navbar(props) {
   const { classes, children } = props;
