+++ src/pages/components/Navbar/styles.js
@@ -1,0 +1,17 @@
+const styles = {
+  root: {
+    flexGrow: 1,
+  },
+  grow: {
+    flexGrow: 1,
+  },
+  menuButton: {
+    marginLeft: -12,
+    marginRight: 20,
+  },
+  loginButton: {
+    marginLeft: '30px',
+  },
+};
+
+export default styles;
